+++ crates/controller-core/src/api/v1/events/mod.rs
@@ -42,6 +42,7 @@ use opentalk_db_storage::{
 };
 use opentalk_keycloak_admin::users::TenantFilter;
 use opentalk_keycloak_admin::KeycloakAdminClient;
+use opentalk_types::core::RoomId;
 use opentalk_types::{
     api::v1::{
         events::{
@@ -1077,9 +1078,89 @@ pub async fn patch_event(
     Ok(Either::Left(ApiResponse::new(event_resource)))
 }
 
-/// Part of `PATCH /events/{event_id}` (see [`patch_event`])
-///
-/// Notify invited users about the event update
+pub async fn notify_event_invitees_by_room_about_update(
+    kc_admin_client: &Data<KeycloakAdminClient>,
+    settings: Arc<Settings>,
+    mail_service: Arc<MailService>,
+    current_tenant: Tenant,
+    current_user: User,
+    conn: &mut DbConnection,
+    room_id: RoomId,
+) -> Result<(), ApiError> {
+    let event = Event::get_first_for_room(conn, room_id).await?;
+
+    if let Some(event) = event {
+        let (event, _invite, room, sip_config, _is_favorite, shared_folder, _tariff) =
+            Event::get_with_related_items(conn, current_user.id, event.id).await?;
+
+        let shared_folder_for_user =
+            shared_folder_for_user(shared_folder, event.created_by, current_user.id);
+
+        notify_event_invitees_about_update(
+            kc_admin_client,
+            settings,
+            mail_service,
+            current_tenant,
+            current_user,
+            conn,
+            event,
+            room,
+            sip_config,
+            shared_folder_for_user,
+        )
+        .await?;
+    }
+    Ok(())
+}
+
+#[allow(clippy::too_many_arguments)]
+async fn notify_event_invitees_about_update(
+    kc_admin_client: &Data<KeycloakAdminClient>,
+    settings: Arc<Settings>,
+    mail_service: Arc<MailService>,
+    current_tenant: Tenant,
+    current_user: User,
+    conn: &mut DbConnection,
+    event: Event,
+    room: Room,
+    sip_config: Option<SipConfig>,
+    shared_folder_for_user: Option<SharedFolder>,
+) -> anyhow::Result<(), ApiError> {
+    let invited_users = get_invited_mail_recipients_for_event(conn, event.id).await?;
+    let current_user_mail_recipient = MailRecipient::Registered(current_user.clone().into());
+    let users_to_notify = invited_users
+        .into_iter()
+        .chain(std::iter::once(current_user_mail_recipient))
+        .collect::<Vec<_>>();
+    let invite_for_room = Invite::get_first_for_room(conn, room.id, current_user.id).await?;
+    let created_by = if event.created_by == current_user.id {
+        current_user
+    } else {
+        User::get(conn, event.created_by).await?
+    };
+
+    let notification_values = UpdateNotificationValues {
+        tenant: current_tenant,
+        created_by,
+        event,
+        event_exception: None,
+        room,
+        sip_config,
+        users_to_notify,
+        invite_for_room,
+    };
+
+    notify_invitees_about_update(
+        settings,
+        notification_values,
+        mail_service,
+        kc_admin_client,
+        shared_folder_for_user,
+    )
+    .await;
+    Ok(())
+}
+
 async fn notify_invitees_about_update(
     settings: Arc<Settings>,
     notification_values: UpdateNotificationValues,